# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Taxi-service",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/operators/assign": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Assign the car to the driver",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operators"
                ],
                "summary": "Assign car to the driver",
                "parameters": [
                    {
                        "description": "Body for assign car",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AssignCarToDriverReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": ""
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    }
                }
            }
        },
        "/operators/car": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Add new car to database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operators"
                ],
                "summary": "Add new car",
                "parameters": [
                    {
                        "description": "Body for add new car request",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddCarReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": ""
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Create a new order as a client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get available orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetAvailableOrdersResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Create a new order as a client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Body for creating new order",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateOrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": ""
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Accept or close the order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Actions with order",
                "parameters": [
                    {
                        "description": "Body for order actions",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OrderActionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": ""
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "AddCarReq": {
            "type": "object",
            "properties": {
                "model": {
                    "type": "string",
                    "example": "BMW"
                },
                "number": {
                    "type": "string",
                    "example": "AX1234XA"
                },
                "series": {
                    "type": "string",
                    "example": "M5"
                },
                "status": {
                    "type": "string",
                    "example": "available"
                }
            }
        },
        "CreateOrderReq": {
            "type": "object",
            "properties": {
                "car_arrival_time": {
                    "type": "string",
                    "example": "2020-11-11T23:30:00Z"
                },
                "destination_point": {
                    "type": "object",
                    "$ref": "#/definitions/Location"
                },
                "starting_point": {
                    "type": "object",
                    "$ref": "#/definitions/Location"
                }
            }
        },
...
        "ErrResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "INTERNAL_SERVER_ERROR"
                }
            }
        },
        "GetAvailableOrdersResp": {
            "type": "object",
            "properties": {
                "car_arrival_time": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "trip_info": {
                    "type": "object",
                    "$ref": "#/definitions/TripInfo"
                }
            }
        },
...
        "Location": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Some City, some Street, 123"
                },
                "destination_point_longitude": {
                    "type": "number",
                    "example": 12.12345
                },
                "latitude": {
                    "type": "number",
                    "example": 12.12345
                }
            }
        },
...
        "TripInfo": {
            "type": "object",
            "properties": {
                "destination_point": {
                    "type": "object",
                    "$ref": "#/definitions/Location"
                },
                "distance": {
                    "type": "number",
                    "example": 15
                },
                "starting_point": {
                    "type": "object",
                    "$ref": "#/definitions/Location"
                }
            }
        },
        "dto.AssignCarToDriverReq": {
            "type": "object",
            "properties": {
                "car_id": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "string"
                }
            }
        },
        "dto.OrderActionsReq": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}